Clean libcore code by splitting into more modules
